requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.11.0
bardapi==0.1.3
googletrans==3.1.0a0
psutil==5.9.5
speedtest-cli==2.1.3
icrawler==0.6.6
